#631 moved DolphinWindow and DolphinStage classes to DolphinWindowUtils
