+++ platform/dolphin-platform-remoting-client-javafx/src/main/java/com/canoo/platform/remoting/client/javafx/window/DolphinWindowUtils.java
@@ -19,6 +19,8 @@ import com.canoo.platform.remoting.client.javafx.view.AbstractViewController;
 import com.canoo.platform.core.functional.Subscription;
 import com.canoo.dp.impl.platform.core.Assert;
 import javafx.event.EventHandler;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
 import javafx.stage.Window;
 import javafx.stage.WindowEvent;
 
@@ -26,6 +28,45 @@ import javafx.stage.WindowEvent;
  * Class that provides some helper methods.
  */
 public class DolphinWindowUtils {
+	
+	/**
+	 * A JavaFX {@link Stage} that contains the view of a {@link AbstractViewController} and will automatically call
+	 * {@link AbstractViewController#destroy()} when the stage becomes hidden.
+	 * @param <M> type of the model
+	 */
+	private static class DolphinStage<M> extends Stage {
+
+	    /**
+	     * Constructor
+	     * @param viewBinder the viewBinder
+	     */
+	    public DolphinStage(final AbstractViewController<M> viewBinder) {
+	        Assert.requireNonNull(viewBinder, "viewBinder");
+	        DolphinWindowUtils.destroyOnClose(this, viewBinder);
+	        setScene(new Scene(viewBinder.getParent()));
+	    }
+	}
+	
+	/**
+	 * A JavaFX {@link Window} that contains the view of a {@link AbstractViewController} and will automatically call
+	 * {@link AbstractViewController#destroy()} when the stage becomes hidden.
+	 *
+	 * @param <M> type of the model
+	 */
+	 private static class DolphinWindow<M> extends Window {
+
+	    /**
+	     * Constructor
+	     * @param viewBinder the viewBinder
+	     */
+	    public DolphinWindow(final AbstractViewController<M> viewBinder) {
+	        Assert.requireNonNull(viewBinder, "viewBinder");
+	        DolphinWindowUtils.destroyOnClose(this, viewBinder);
+	        setScene(new Scene(viewBinder.getParent()));
+	    }
+	    
+	   
+	}
 
     /**
      * The method will register an event handler to the window that will automatically call the {@link AbstractViewController#destroy()}
@@ -42,5 +83,26 @@ public class DolphinWindowUtils {
         window.addEventFilter(WindowEvent.WINDOW_HIDDEN, handler);
         return () -> window.removeEventFilter(WindowEvent.WINDOW_HIDDEN, handler);
     }
+
+	/**
+	 * The method will create a JavaFX {@link Window} that contains the view of a {@link AbstractViewController} and will automatically call
+	 * {@link AbstractViewController#destroy()} when the stage becomes hidden.
+	 *
+	 * @param <M> type of the model
+	 */
+    public static final <M> Window createWindow(AbstractViewController<M> viewBinder) {
+       Assert.requireNonNull(viewBinder, "viewBinder");
+	   return new DolphinWindow<M>(viewBinder);
+    }
+    
+	/**
+	 * The method will create a JavaFX {@link Stage} that contains the view of a {@link AbstractViewController} and will automatically
+	 * call {@link AbstractViewController#destroy()} when the stage becomes hidden.
+	 * @param <M> type of the model
+	 */
+    public static final <M> Stage createStage(final AbstractViewController<M> viewBinder) {
+    	Assert.requireNonNull(viewBinder, "viewBinder");
+    	return new DolphinStage<M>(viewBinder);
+     }
 
 }
